moved deliveryservice_regexes, profile, and type structs out to lib/go-tc
