+++ lib/go-tc/deliveryservice_ssl_keys.go
@@ -1,0 +1,42 @@
+package tc
+
+/*
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+// DeliveryServiceSSLKeysResponse ...
+type DeliveryServiceSSLKeysResponse struct {
+	Response DeliveryServiceSSLKeys `json:"response"`
+}
+
+// DeliveryServiceSSLKeys ...
+type DeliveryServiceSSLKeys struct {
+	CDN             string                            `json:"cdn"`
+	DeliveryService string                            `json:"DeliveryService"`
+	BusinessUnit    string                            `json:"businessUnit"`
+	City            string                            `json:"city"`
+	Organization    string                            `json:"organization"`
+	Hostname        string                            `json:"hostname"`
+	Country         string                            `json:"country"`
+	State           string                            `json:"state"`
+	Version         string                            `json:"version"`
+	Certificate     DeliveryServiceSSLKeysCertificate `json:"certificate"`
+}
+
+// DeliveryServiceSSLKeysCertificate ...
+type DeliveryServiceSSLKeysCertificate struct {
+	Crt string `json:"crt"`
+	Key string `json:"key"`
+	CSR string `json:"csr"`
+}
