+++ traffic_ops/client/delivery_service.go
@@ -34,8 +34,8 @@ func (to *Session) DeliveryServices() ([]tc.DeliveryService, error) {
 }
 
 // DeliveryServices gets an array of DeliveryServices
-func (to *Session) DeliveryServicesByServer(id int) ([]DeliveryService, error) {
-	var data GetDeliveryServiceResponse
+func (to *Session) DeliveryServicesByServer(id int) ([]tc.DeliveryService, error) {
+	var data tc.GetDeliveryServiceResponse
 	err := get(to, deliveryServicesByServerEp(strconv.Itoa(id)), &data)
 	if err != nil {
 		return nil, err
@@ -45,8 +45,8 @@ func (to *Session) DeliveryServicesByServer(id int) ([]DeliveryService, error) {
 }
 
 // DeliveryService gets the DeliveryService for the ID it's passed
-func (to *Session) DeliveryService(id string) (*DeliveryService, error) {
-	var data GetDeliveryServiceResponse
+func (to *Session) DeliveryService(id string) (*tc.DeliveryService, error) {
+	var data tc.GetDeliveryServiceResponse
 	err := get(to, deliveryServiceEp(id), &data)
 	if err != nil {
 		return nil, err
@@ -56,8 +56,8 @@ func (to *Session) DeliveryService(id string) (*DeliveryService, error) {
 }
 
 // CreateDeliveryService creates the DeliveryService it's passed
-func (to *Session) CreateDeliveryService(ds *DeliveryService) (*CreateDeliveryServiceResponse, error) {
-	var data CreateDeliveryServiceResponse
+func (to *Session) CreateDeliveryService(ds *tc.DeliveryService) (*tc.CreateDeliveryServiceResponse, error) {
+	var data tc.CreateDeliveryServiceResponse
 	jsonReq, err := json.Marshal(ds)
 	if err != nil {
 		return nil, err
@@ -72,8 +72,8 @@ func (to *Session) CreateDeliveryService(ds *DeliveryService) (*CreateDeliverySe
 
 // UpdateDeliveryService updates the DeliveryService matching the ID it's passed with
 // the DeliveryService it is passed
-func (to *Session) UpdateDeliveryService(id string, ds *DeliveryService) (*UpdateDeliveryServiceResponse, error) {
-	var data UpdateDeliveryServiceResponse
+func (to *Session) UpdateDeliveryService(id string, ds *tc.DeliveryService) (*tc.UpdateDeliveryServiceResponse, error) {
+	var data tc.UpdateDeliveryServiceResponse
 	jsonReq, err := json.Marshal(ds)
 	if err != nil {
 		return nil, err
@@ -87,8 +87,8 @@ func (to *Session) UpdateDeliveryService(id string, ds *DeliveryService) (*Updat
 }
 
 // DeleteDeliveryService deletes the DeliveryService matching the ID it's passed
-func (to *Session) DeleteDeliveryService(id string) (*DeleteDeliveryServiceResponse, error) {
-	var data DeleteDeliveryServiceResponse
+func (to *Session) DeleteDeliveryService(id string) (*tc.DeleteDeliveryServiceResponse, error) {
+	var data tc.DeleteDeliveryServiceResponse
 	err := del(to, deliveryServiceEp(id), &data)
 	if err != nil {
 		return nil, err
@@ -98,8 +98,8 @@ func (to *Session) DeleteDeliveryService(id string) (*DeleteDeliveryServiceRespo
 }
 
 // DeliveryServiceState gets the DeliveryServiceState for the ID it's passed
-func (to *Session) DeliveryServiceState(id string) (*DeliveryServiceState, error) {
-	var data DeliveryServiceStateResponse
+func (to *Session) DeliveryServiceState(id string) (*tc.DeliveryServiceState, error) {
+	var data tc.DeliveryServiceStateResponse
 	err := get(to, deliveryServiceStateEp(id), &data)
 	if err != nil {
 		return nil, err
@@ -109,8 +109,8 @@ func (to *Session) DeliveryServiceState(id string) (*DeliveryServiceState, error
 }
 
 // DeliveryServiceHealth gets the DeliveryServiceHealth for the ID it's passed
-func (to *Session) DeliveryServiceHealth(id string) (*DeliveryServiceHealth, error) {
-	var data DeliveryServiceHealthResponse
+func (to *Session) DeliveryServiceHealth(id string) (*tc.DeliveryServiceHealth, error) {
+	var data tc.DeliveryServiceHealthResponse
 	err := get(to, deliveryServiceHealthEp(id), &data)
 	if err != nil {
 		return nil, err
@@ -120,8 +120,8 @@ func (to *Session) DeliveryServiceHealth(id string) (*DeliveryServiceHealth, err
 }
 
 // DeliveryServiceCapacity gets the DeliveryServiceCapacity for the ID it's passed
-func (to *Session) DeliveryServiceCapacity(id string) (*DeliveryServiceCapacity, error) {
-	var data DeliveryServiceCapacityResponse
+func (to *Session) DeliveryServiceCapacity(id string) (*tc.DeliveryServiceCapacity, error) {
+	var data tc.DeliveryServiceCapacityResponse
 	err := get(to, deliveryServiceCapacityEp(id), &data)
 	if err != nil {
 		return nil, err
@@ -131,8 +131,8 @@ func (to *Session) DeliveryServiceCapacity(id string) (*DeliveryServiceCapacity,
 }
 
 // DeliveryServiceRouting gets the DeliveryServiceRouting for the ID it's passed
-func (to *Session) DeliveryServiceRouting(id string) (*DeliveryServiceRouting, error) {
-	var data DeliveryServiceRoutingResponse
+func (to *Session) DeliveryServiceRouting(id string) (*tc.DeliveryServiceRouting, error) {
+	var data tc.DeliveryServiceRoutingResponse
 	err := get(to, deliveryServiceRoutingEp(id), &data)
 	if err != nil {
 		return nil, err
@@ -142,8 +142,8 @@ func (to *Session) DeliveryServiceRouting(id string) (*DeliveryServiceRouting, e
 }
 
 // DeliveryServiceServer gets the DeliveryServiceServer
-func (to *Session) DeliveryServiceServer(page, limit string) ([]DeliveryServiceServer, error) {
-	var data DeliveryServiceServerResponse
+func (to *Session) DeliveryServiceServer(page, limit string) ([]tc.DeliveryServiceServer, error) {
+	var data tc.DeliveryServiceServerResponse
 	err := get(to, deliveryServiceServerEp(page, limit), &data)
 	if err != nil {
 		return nil, err
@@ -153,8 +153,8 @@ func (to *Session) DeliveryServiceServer(page, limit string) ([]DeliveryServiceS
 }
 
 // DeliveryServiceServer gets the DeliveryServiceServer
-func (to *Session) DeliveryServiceRegexes() ([]DeliveryServiceRegexes, error) {
-	var data DeliveryServiceRegexResponse
+func (to *Session) DeliveryServiceRegexes() ([]tc.DeliveryServiceRegexes, error) {
+	var data tc.DeliveryServiceRegexResponse
 	err := get(to, deliveryServiceRegexesEp(), &data)
 	if err != nil {
 		return nil, err
@@ -164,8 +164,8 @@ func (to *Session) DeliveryServiceRegexes() ([]DeliveryServiceRegexes, error) {
 }
 
 // DeliveryServiceSSLKeysByID gets the DeliveryServiceSSLKeys by ID
-func (to *Session) DeliveryServiceSSLKeysByID(id string) (*DeliveryServiceSSLKeys, error) {
-	var data DeliveryServiceSSLKeysResponse
+func (to *Session) DeliveryServiceSSLKeysByID(id string) (*tc.DeliveryServiceSSLKeys, error) {
+	var data tc.DeliveryServiceSSLKeysResponse
 	err := get(to, deliveryServiceSSLKeysByIDEp(id), &data)
 	if err != nil {
 		return nil, err
@@ -175,8 +175,8 @@ func (to *Session) DeliveryServiceSSLKeysByID(id string) (*DeliveryServiceSSLKey
 }
 
 // DeliveryServiceSSLKeysByHostname gets the DeliveryServiceSSLKeys by Hostname
-func (to *Session) DeliveryServiceSSLKeysByHostname(hostname string) (*DeliveryServiceSSLKeys, error) {
-	var data DeliveryServiceSSLKeysResponse
+func (to *Session) DeliveryServiceSSLKeysByHostname(hostname string) (*tc.DeliveryServiceSSLKeys, error) {
+	var data tc.DeliveryServiceSSLKeysResponse
 	err := get(to, deliveryServiceSSLKeysByHostnameEp(hostname), &data)
 	if err != nil {
 		return nil, err
